refactor: 2차 문지기를 DB atomic UPDATE에서 INSERT-only 방식으로 전환

+++ apps/commerce-api/src/main/java/com/loopers/interfaces/coupon/dto/AdminCouponV1Dto.java
@@ -52,7 +52,7 @@ public class AdminCouponV1Dto {
         long discountValue,
         Long minOrderAmount,
         int totalQuantity,
-        int issuedQuantity,
+        long issuedQuantity,
         ZonedDateTime expiredAt,
         ZonedDateTime createdAt,
         ZonedDateTime updatedAt
@@ -72,7 +72,7 @@ public class AdminCouponV1Dto {
         String discountType,
         long discountValue,
         int totalQuantity,
-        int issuedQuantity,
+        long issuedQuantity,
         ZonedDateTime expiredAt,
         ZonedDateTime createdAt
     ) {
